Arrumado edição do usuário, codigo estava quebrado, arrumado cadastro de hora onde o mesmo estavca dando página whitelabel, feito ajustes nas entidades "pessoa, usuário".

+++ aplicacao/src/main/java/com/tcc/aplicacao/entities/BancoHoras.java
@@ -16,11 +16,30 @@ public class BancoHoras {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
-
-    private Double saldoMensal;
-    private Double saldoAtual;
+    private long saldoMensal;
+    private long saldoAtual;
     private Boolean saldoNegativo;
 
     @ManyToOne(optional = true)
     private Usuario usuario;
 }
+
+// -- Criação da função para atualizar saldo negativo ou positivo
+
+// CREATE OR REPLACE FUNCTION atualiza_saldo_negativo()
+// RETURNS TRIGGER AS $$
+// BEGIN
+// IF NEW.saldo_atual < NEW.saldo_mensal THEN
+// NEW.saldo_negativo := TRUE;
+// ELSE
+// NEW.saldo_negativo := FALSE;
+// END IF;
+// RETURN NEW;
+// END;
+// $$ LANGUAGE plpgsql;
+
+// -- Criação do trigger
+// CREATE TRIGGER trigger_atualiza_saldo_negativo
+// BEFORE INSERT OR UPDATE ON public.banco_horas
+// FOR EACH ROW
+// EXECUTE FUNCTION atualiza_saldo_negativo();
